+++ examples/common/client-demo/src/BacnetDemoLogging.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include <EspBacnet.h>
+#include <BacnetClient.h>
 
 #include "ConfigManager.h"
 #include "BacnetDemoWatchedAnalogValue.h"